--- ciqual/ssconvert.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-ssconvert -S Ciqual.xls Ciqual.csv
